+++ src/tree/Clustering.h
@@ -4,7 +4,12 @@
 
 class IClustering {
 public:
-	virtual void operator()(const float* distanceMatrix, size_t n_elems, size_t n_medoids, size_t n_fixed_medoids, int* centers) = 0;
+	virtual void operator()(
+		const float* distanceMatrix, 
+		int n_elems, 
+		int n_medoids, 
+		int n_fixed_medoids, 
+		int* centers) = 0;
 };
 
 class CLARANS : public IClustering {
@@ -17,16 +22,16 @@ public:
 
 	CLARANS(float exploreFraction, int numLocal) : exploreFraction(exploreFraction), numLocal(numLocal) {}
 
-	void operator()(const float* distanceMatrix, size_t n_elems, size_t n_medoids, size_t n_fixed_medoids, int* medoids) override;
+	void operator()(const float* distanceMatrix, int n_elems, int n_medoids, int n_fixed_medoids, int* medoids) override;
 
 protected:
 
-	float calculateCost(const float* distanceMatrix, int *candidate, size_t n_elems, size_t n_medoids);
+	float calculateCost(const float* distanceMatrix, int *candidate, int n_elems, int n_medoids);
 
 	void updateAssignment(
 		int x,
 		int *candidate,
-		size_t n_medoids,
+		int n_medoids,
 		const float* D,
 		float& dist_nearest,
 		float& dist_second,
