+++ src/tree/GuideTree.cpp
@@ -63,14 +63,14 @@ bool GuideTree::saveNewick(
 
 
 // *******************************************************************
-uint64_t GuideTree::calculateSackinIndex() {
+int64_t GuideTree::calculateSackinIndex() {
 	
 	uint64_t idx = 0;
-	size_t n_sequences = getSequenceCount();
+	int n_sequences = getSequenceCount();
 	
 	if (n_sequences) {
-		std::vector<size_t> depths(this->guide_tree.size());
-		for (size_t i = guide_tree.size() - 1; i >= n_sequences; --i)
+		std::vector<int64_t> depths(this->guide_tree.size());
+		for (int i = (int)guide_tree.size() - 1; i >= n_sequences; --i)
 		{
 			depths[guide_tree[i].first] = depths[i] + 1;
 			depths[guide_tree[i].second] = depths[i] + 1;
