+++ src/utils/log.cpp
@@ -88,7 +88,7 @@ std::string Log::formatLargeNumber(uint64_t num, int minWidth) {
 
 	} while (num > 0);
 
-	int initialSpaces = minWidth - out.length();
+	int initialSpaces = (int) (minWidth - out.length());
 
 	if (initialSpaces > 0) {
 		out = string(initialSpaces, ' ') + out;
